+++ calcipy/noxfile/__init__.py
@@ -1,6 +1,6 @@
 try:
-    from ._noxfile import build_check, build_dist, tests
+    from ._noxfile import tests
 except ImportError as exc:
     raise RuntimeError("The 'calcipy[nox]' extras are missing") from exc
 
-__all__ = ('build_check', 'build_dist', 'tests')
+__all__ = ('tests',)